refactor dropdown clickable and positioning
