+++ src/components/Global/DropdownSearch/DropdownSearch.module.css
@@ -4,41 +4,31 @@
   flex-direction: column;
   border-radius: 4px;
   position: relative;
-  display: inline-block;
   
   
-}
-.blur_container{
-  background-color: red;
-  position: absolute;
-  width: 800px;
-  height: 500px;
-  top: 100%;
-  left: 0;
-  z-index: 1;
 }
 
 
 
 .dropdown_container{
 
-    position: absolute;
-    display: flex;
+  position: absolute;
+  display: flex;
 
 padding: 16px;
 flex-direction: column;
 border-radius: 4px;
 border: 1px solid var(--dark3, #1E1E24);
 background: var(--dark4, #0E0E14);
-box-shadow: 0px 20px 20px 0px rgba(0, 0, 0, 0.25);
+/* box-shadow: 0px 20px 20px 0px rgba(0, 0, 0, 0.25); */
 box-shadow: rgba(115, 113, 252, 0.56) 0px 100px 100px 2px;
 gap: 8px;
-  top: 100%;
-  left: 0;
-  z-index: 2;
+top: 30px;
+left: 0;
+z-index: 30;
 
 }
- 
+
  .dropdown_content{
     display: flex;
 width: 515px;
