+++ src/pages/Trade/TradeCharts/TradeChartsComponents/TradeChartsTokenInfo.tsx
@@ -1,15 +1,11 @@
 import React, { memo, useContext } from 'react';
-import useMediaQuery from '../../../../utils/hooks/useMediaQuery';
 
 import { PoolContext } from '../../../../contexts/PoolContext';
-import TokenIcon from '../../../../components/Global/TokenIcon/TokenIcon';
 import {
     getFormattedNumber,
     getUnicodeCharacter,
 } from '../../../../ambient-utils/dataLayer';
-import { TokenIF } from '../../../../ambient-utils/types';
 import { FlexContainer } from '../../../../styled/Common';
-import { HeaderButtons, HeaderText } from '../../../../styled/Components/Chart';
 import { TradeDataContext } from '../../../../contexts/TradeDataContext';
 import DropdownSearch from '../../../../components/Global/DropdownSearch/DropdownSearch';
 import PoolData from './PoolData';
@@ -28,10 +24,6 @@ function TradeChartsTokenInfo() {
 
     const denomInBase = isDenomBase;
 
-    const [topToken, bottomToken]: [TokenIF, TokenIF] = denomInBase
-        ? [baseToken, quoteToken]
-        : [quoteToken, baseToken];
-
     const currencyCharacter = denomInBase
         ? // denom in a, return token b character
           getUnicodeCharacter(quoteToken.symbol)
@@ -49,8 +41,6 @@ function TradeChartsTokenInfo() {
         abbrevThreshold: 10000000, // use 'm', 'b' format > 10m
     });
 
-    const smallScrenView = useMediaQuery('(max-width: 968px)');
-
     const poolPrice = isTradeDollarizationEnabled
         ? usdPrice
             ? getFormattedNumber({ value: usdPrice, prefix: '$' })
@@ -64,48 +54,6 @@ function TradeChartsTokenInfo() {
     const poolPriceChangeString =
         poolPriceChangePercent === undefined ? '…' : poolPriceChangePercent;
 
-    const denomToggleButton = (
-        <FlexContainer gap={8}>
-            <HeaderButtons
-                id='token_pair_in_chart_header'
-                aria-label='flip denomination.'
-                onClick={() => toggleDidUserFlipDenom()}
-            >
-                <FlexContainer
-                    id='trade_chart_header_token_pair_logos'
-                    role='button'
-                    gap={8}
-                >
-                    <TokenIcon
-                        token={topToken}
-                        src={topToken.logoURI}
-                        alt={topToken.symbol}
-                        size={smallScrenView ? 's' : 'l'}
-                    />
-                    <TokenIcon
-                        token={bottomToken}
-                        src={bottomToken.logoURI}
-                        alt={bottomToken.symbol}
-                        size={smallScrenView ? 's' : 'l'}
-                    />
-                </FlexContainer>
-                <HeaderText
-                    id='trade_chart_header_token_pair_symbols'
-                    fontSize='header1'
-                    fontWeight='300'
-                    color='text1'
-                    role='button'
-                    aria-live='polite'
-                    aria-atomic='true'
-                    aria-relevant='all'
-                >
-                    {topToken.symbol} / {bottomToken.symbol}
-                </HeaderText>
-            </HeaderButtons>
-            <DropdownSearch />
-        </FlexContainer>
-    );
-
     const poolDataProps = {
         poolPrice,
         poolPriceChangeString,
@@ -114,7 +62,7 @@ function TradeChartsTokenInfo() {
     };
     return (
         <FlexContainer alignItems='center' gap={16}>
-            {denomToggleButton}
+            <DropdownSearch />
 
             <PoolData {...poolDataProps} />
         </FlexContainer>
